조직 설계 화면 노출시 scope 1, 2, 3을 수정하도록 수정

+++ Carbon/views.py
@@ -1,4 +1,5 @@
 import json
+import datetime
 
 from rest_framework.views import APIView
 from rest_framework.response import Response
@@ -21,6 +22,8 @@ from Swag import CarSwag
 class CarbonEmissionQuery(APIView):
 
     permission_classes = (IsAuthenticated,)  # 로그인 검증
+
+    year = datetime.date.__str__(datetime.date.today())[:4]
 
     @swagger_auto_schema(
         operation_summary="요청한 회사의 모든 탄소 배출원을 반환하는 Api",
@@ -187,25 +190,33 @@ class CarbonEmissionQuery(APIView):
         except KeyError:
             return Response("Wrong CarbonInfo Data", status=status.HTTP_410_GONE)
 
-        if CarInfoTemp.Scope == 1:
-            TargetCom.Scope1 += round(CarTrans, 4)
-        elif CarInfoTemp.Scope == 2:
-            TargetCom.Scope2 += round(CarTrans, 4)
-        elif CarInfoTemp.Scope == 3:
-            TargetCom.Scope3 += round(CarTrans, 4)
-        else:
-            CarInfoTemp.delete()
-            return Response(
-                "Wrong Scope Num Entered", status=status.HTTP_406_NOT_ACCEPTABLE
-            )
-
-        TargetCom.save()
-
         try:
             if type(TargetCom) == ComModel.Company:
                 func.CreateCarbon(
                     CarbonData, round(CarTrans, 4), usage, UserRoot, None, CarInfoTemp
                 )
+
+                if int(self.year) >= int(CarInfoTemp.EndDate[:4]):
+                    if CarInfoTemp.Scope == 1:
+                        TargetCom.Scope1 += func.DivideByMonthOrYear(
+                            CarInfoTemp.StartDate, CarInfoTemp.EndDate, CarTrans, 1
+                        )
+                    elif CarInfoTemp.Scope == 2:
+                        TargetCom.Scope2 += func.DivideByMonthOrYear(
+                            CarInfoTemp.StartDate, CarInfoTemp.EndDate, CarTrans, 1
+                        )
+                    elif CarInfoTemp.Scope == 3:
+                        TargetCom.Scope3 += func.DivideByMonthOrYear(
+                            CarInfoTemp.StartDate, CarInfoTemp.EndDate, CarTrans, 1
+                        )
+                    else:
+                        CarInfoTemp.delete()
+                        return Response(
+                            "Wrong Scope Num Entered",
+                            status=status.HTTP_406_NOT_ACCEPTABLE,
+                        )
+
+                    TargetCom.save()
 
             else:
                 func.CreateCarbon(
@@ -216,6 +227,29 @@ class CarbonEmissionQuery(APIView):
                     TargetCom,
                     CarInfoTemp,
                 )
+
+                if int(self.year) >= int(CarInfoTemp.EndDate[:4]):
+
+                    if CarInfoTemp.Scope == 1:
+                        TargetCom.SelfCom.Scope1 += func.DivideByMonthOrYear(
+                            CarInfoTemp.StartDate, CarInfoTemp.EndDate, CarTrans, 1
+                        )
+                    elif CarInfoTemp.Scope == 2:
+                        TargetCom.Scope2 += func.DivideByMonthOrYear(
+                            CarInfoTemp.StartDate, CarInfoTemp.EndDate, CarTrans, 1
+                        )
+                    elif CarInfoTemp.Scope == 3:
+                        TargetCom.SelfCom.Scope3 += func.DivideByMonthOrYear(
+                            CarInfoTemp.StartDate, CarInfoTemp.EndDate, CarTrans, 1
+                        )
+                    else:
+                        CarInfoTemp.delete()
+                        return Response(
+                            "Wrong Scope Num Entered",
+                            status=status.HTTP_406_NOT_ACCEPTABLE,
+                        )
+
+                    TargetCom.SelfCom.save()
 
         except KeyError:
             return Response("Wrong Carbon Data", status=status.HTTP_410_GONE)
